drm/imagination: Peg the selected device page size to the CPU page size
Most of the heavy lifting is already done by %PVR_DEVICE_PAGE_SIZE, so
we just define it to be %PAGE_SIZE. The remaining variations are
determined by a lump of macros at the top of pvr_vm.c, which also
verifies at build time whether the selected CPU page size is also a
valid device page size. If this check fails, so does the build.

Signed-off-by: Matt Coster <matt.coster@imgtec.com>
Signed-off-by: Frank Binns <frank.binns@imgtec.com>